Enums elements were added.

+++ modules/orm/src/main/java/nc/sumy/edu/webapp/orm/domain/User.java
@@ -25,7 +25,7 @@ public class User {
         this.birthDate = birthDate;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
